Task 1 - Data redundancy/backend1/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.3.0
